requirements.txt
pytest
pytest-cov
numpy
pandas>=1.5.0
scikit-learn==1.8.0
imbalanced-learn==0.14.0
joblib==1.5.2
matplotlib>=3.0.0
seaborn
Pillow>=9.0.0
playwright>=1.30.0
beautifulsoup4==4.14.3
lxml>=4.9.0
requests>=2.28.0
cssselect==1.3.0
html5lib
click==8.3.1
PyYAML>=6.0
rich>=13.0.0
tqdm
regex>=2022.10.31
argcomplete==3.1.4
attrs==23.2.0
colorama==0.4.6
coverage==7.13.0
dill==0.4.0
filelock==3.13.1
idna==3.6
Jinja2==3.1.2
jsonschema==4.10.3
packaging